Combine two servlet files and reduce round-trips to the server
Delete the servlet that handles DELETE request and combine the two servlet files into one DataServlet.java file at the "/comments" URL that handles GET, POST and DELETE request. Update the index.html file to accomodate the change.
Reduce the round-trips to the server by modifying the onclick function of the buttons and let the buttons take in the string format of the entity key instead of the comment itself.
Modify the LikeDataServlet.java file to set error status code to 404 when the entity is not found.
