+++ portfolio/src/main/webapp/index.html
@@ -58,7 +58,7 @@
       </div>
       <!-- Play around with Java Servlets and try to let the client user enter their name -->
       <h2 id="comments">Leave your comments</h2>
-      <form action="/data" method="POST">
+      <form action="/comments" method="POST">
         <p>Feel free to leave your comments here:</p>
         <textarea name="comment-input" placeholder="Your Comments"></textarea>
         <br/><br/>
